MYR-870: move loading to component
