+++ src/components-v2/atoms/Loading/index.ts
@@ -1,0 +1,1 @@
+export * from './Loading';
